+++ blog/static/babel.config.cjs
@@ -1,0 +1,12 @@
+module.exports = {
+  presets: [
+    [
+      "@babel/preset-env",
+      {
+        targets: {
+          node: "current", // Ensure Babel transpiles for your current Node version.
+        },
+      },
+    ],
+  ],
+};
