+++ blog/templates/blog/base.html
@@ -67,7 +67,7 @@
             <img src="{% static 'images/jeezus.jpg' %}" alt="The Christ" id="jesusimg">
         </footer>
         <script async defer src="https://dailyverses.net/get/random.js?language=esv"></script>
-        <script src="{% static 'js/scripts.js' %}"></script>
+        <script type="module" src="{% static 'js/scripts.js' %}"></script>
         <script src="{% static 'js/splash.js' %}"></script>
     </body>
 </html>
